[refactor] externalize config helpers to make them testables
